Remove the WaitCacheSync block on informers
The WaitForCacheSync waits forever and never returns in the case a
persisten error occurs. On the other hand, it looks like there is no
way in the current version of surfacing informer problems to the caller,
as stated in this issue: https://github.com/kubernetes/client-go/issues/155.
Error handling for informers has been added recently in
https://github.com/kubernetes/kubernetes/pull/87329 which is only available
in master for the moment.