Improve status reconciliation for EnvoyConfigRevision controller
Reconciliation of the status was performed based on a certain sequantiality of the states an EnvoyConfigRevision passed since creation until publication. This approach does not allow to know the exact status of the revision, as it is based on events occurring in a certain order and not in observations of the real status.

This commit changes that so the status clearly reflects if the resources in the revision are correctly synced with the xDS server cache or not. It does so using the only possible way, inspecting the contents of the xDS server cache.