Drop usage of ManageError and ManageSuccess functions
ManageError(...) and ManageSuccess(...) from https://github.com/redhat-cop/operator-utils have functionality to manage the status of the custom resource after each reconcile, either for success or for error situations. We are not currently using this functionality as the status of marin3r custom resources does not support this.

Furthermore, ManageError is causing some spurious panics in reconcile loops when the custom resource cannot be retrieved for some reason. A first look into ManageError seems to point to the problem being that ManageError only restrains itslef from trying to access the fields in the custom resource if the api call returned a "not found" error. The panic occurs when a different api/client error is returned and then the function tries to access the struct, which just holds a nil value. The statement that ends up causing the panic is https://github.com/redhat-cop/operator-utils/blob/eca9bb2b140c4a08a13a5c570d5cdbcaa15a48b3/pkg/util/lockedresourcecontroller/resource-reconciler.go#L233