Generate bootstrap config in an init container
Generating the boostrap config in an init container when each envoy Pod
starts allows to inject meradata in the envoy node information that will
then be sent to the xDS server. This will allow to keep track of
information like the pod name associated with a request or a failure.

To implemente this, usaage of EnvoyBootstrap resource in the EnvoyDeploymeny
is dropped and a new init container added which calls the new
"init-manager" command that will generate the config and also add some
pod metadata to it.