Change to a per type URL versioning of resources
To avoid unnecessary updates use different versions for each type of
resources. Using the same version for all sends unnecessary updates to
ohe envoy clients which can have a performace penalty.
Add a new field to the EnvoyConfigRevision resource to track the
versions of the envoy resources it provides to the xDS server.

Disable the self-healing feature temporarily as a reimplementation needs
to be done after this change.