Implement a backoff strategy for failure retries
When a NACK is received by the discovery service, it immediately retries
the the same response. This usually does not help because unless
something changes in the resource, the client is likely to reject the
resource again and again. This leads to a storm of retries that elevates
CPU comsumption in the clients, even triggering HPA if enabled.
This commit implements a backoff strategy for retries to avoid this
situation.