Enable multinamespaced mode for the webhook
Since now the bootstrap config for the envoy pods is generated by an
init container, the webhook needs to inject the address of the discovery
service in the Pod. To do so it inspects the DiscoveryService resources
present in the Pod's namespace. This was failing for the multinamespaced
OLM install mode as the code to manage this was missing from the
webhook.