Use the built image as the default for Envoy init and shutdown
While the Envoy init and shutdown manager images can be overridden at
runtime, the defaults were more-or-less hardcoded. Now the default
image is the same as what was built. This means that the init manager
and shutdown manager images will respect build-time overrides of
IMAGE_TAG_BASE, e.g., to use a different registry.

NOTE: I wanted to query at runtime the image that the manager
container is currently running, but that turns out to be difficult to
do. The image name isn't available in the downstream API[1] so for now
I'm doing a cut-n-paste of the existing "version" code and having the
Makefile pass the image name into the generator.

[1] https://github.com/kubernetes/kubernetes/issues/19475