Rework the SnapshotCache
The test of different packages access the internal structure of the
snapshot cache to validate behavior. This causes a lot of work each time
the internal structure of the snapshot cache changes in upstream
go-control-plane. This commit reworks the cache to make use of new
funcions in go-control-plane that make it easier to write/read to/from
the cache and confines any access to the internal structure to the
pkg/discoveryservice/xdss/v3 package.